Metadata is now a struct instead of a class
The various elements of metadata are now directly accessible to the
user. In theory, the user could change the metadata contents causing bad
things to happen. The alternative is to provide getter methods for each
and every bit of metadata. This could be a goal for a future release.
For now, direct access solves a lot of immediate problems and this makes
it worth the risk.