Implement reading an entire frame
When a frame is requested, the entire frame is read at once and stored
in a vector. Since the sizeof() a vector is always 3 bytes, this
necessitates changes to the retrieve() function where the size of the
contents of the vector can be specified. This can be more elegantly
handled by trying value.size() and if that fails, falling back to
sizeof( value ). For now this approach will do.