Demo now writes out the average frame as a data file
The demo has now reached feature parity with release 1.0. When run, it
finds all files in the current directory and writes out some metadata
about the files found. It also demonstrates how to extract a single
pixel (not very useful) or a single frame (much more useful). Finally,
it demonstrates how to get the average of all frames and write that to a
data file.