Move to template functions for reading data
The datatype used to store information in the SPE file controlled how
many bytes are read from the file. This was implemented in a rather
hackish way by having a scoped variable declared in different types
under different conditions and having a lot of code that was repeated
unchanged.

All this is now implemented as a template function that reduces a lot of
code repetition and make the code easier to read.